Helpers: Add `enum Error`
Error type to be used in-between api and rpc-services
