+++ helpers/Cargo.toml
@@ -10,3 +10,4 @@ edition = "2018"
 diesel = { version = "1.4", features = ["postgres", "r2d2"] }
 diesel_migrations = "1.4"
 log = "0.4"
+serde = { version = "1.0", features = ["derive"] }
