+++ helpers/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod db;
+pub mod rpc;
